rem_feature is added in the edit room

+++ admin/ajax/features_facilities.php
@@ -41,10 +41,18 @@
     {
         $frm_data = filteration($_POST);
         $values = [$frm_data['rem_feature']];
+
+        $check_q = select('SELECT * FROM `room_features` WHERE `features_id`=?',[$frm_data['rem_feature']],'i');
+
+        if(mysqli_num_rows($check_q)==0){
+            $q = "DELETE FROM `features` WHERE `id`=?";
+            $res = delete($q,$values,'i'); 
+            echo $res;
+        }
+        else{
+            echo 'room_added';
+        }
         
-        $q = "DELETE FROM `features` WHERE `id`=?";
-        $res = delete($q,$values,'i'); 
-        echo $res;
     }
     
     if (isset($_POST['add_facility'])) {
